requirements.txt
python-dateutil
requests
selectorlib
numpy
boto3
pandas
streamlit